Move listing detail-building into default get_host_and_agent_details implementation
Refactors _collect_and_emit_details_for_host to delegate all host/agent
detail building to ProviderInstanceInterface.get_host_and_agent_details,
which now has a default implementation instead of returning None. This
removes the "optimized vs fallback" bifurcation in the listing pipeline.

The Modal provider overrides get_host_and_agent_details for its SSH-optimized
path and calls super() when falling back to offline data or on connection errors.
