+++ libs/mng/imbue/mng/interfaces/provider_instance.py
@@ -1,7 +1,11 @@
 from abc import ABC
 from abc import abstractmethod
+from collections.abc import Callable
 from concurrent.futures import Future
+from datetime import datetime
+from datetime import timezone
 from pathlib import Path
+from typing import Any
 from typing import Mapping
 from typing import Sequence
 
@@ -13,6 +17,10 @@ from imbue.concurrency_group.concurrency_group import ConcurrencyGroup
 from imbue.concurrency_group.executor import ConcurrencyGroupExecutor
 from imbue.imbue_common.mutable_model import MutableModel
 from imbue.mng.config.data_types import MngContext
+from imbue.mng.errors import HostAuthenticationError
+from imbue.mng.errors import HostConnectionError
+from imbue.mng.hosts.common import compute_idle_seconds
+from imbue.mng.interfaces.agent import AgentInterface
 from imbue.mng.interfaces.data_types import AgentDetails
 from imbue.mng.interfaces.data_types import HostDetails
 from imbue.mng.interfaces.data_types import HostLifecycleOptions
@@ -22,18 +30,181 @@ from imbue.mng.interfaces.data_types import VolumeInfo
 from imbue.mng.interfaces.host import HostInterface
 from imbue.mng.interfaces.host import OnlineHostInterface
 from imbue.mng.interfaces.volume import HostVolume
+from imbue.mng.primitives import ActivitySource
 from imbue.mng.primitives import AgentId
+from imbue.mng.primitives import AgentLifecycleState
+from imbue.mng.primitives import CommandString
 from imbue.mng.primitives import DiscoveredAgent
 from imbue.mng.primitives import DiscoveredHost
 from imbue.mng.primitives import HostId
 from imbue.mng.primitives import HostName
 from imbue.mng.primitives import HostNameStyle
+from imbue.mng.primitives import HostState
 from imbue.mng.primitives import ImageReference
 from imbue.mng.primitives import ProviderInstanceName
+from imbue.mng.primitives import SSHInfo
 from imbue.mng.primitives import SnapshotId
 from imbue.mng.primitives import SnapshotName
 from imbue.mng.primitives import VolumeId
 from imbue.mng.utils.name_generator import generate_host_name
+
+
+def _build_host_details_from_host(
+    host: HostInterface,
+    host_ref: DiscoveredHost,
+    is_authentication_failure: bool,
+) -> tuple[HostDetails, datetime | None]:
+    """Build HostDetails from a host object (online or offline).
+
+    Returns the HostDetails and the SSH activity time (needed for agent idle calculation).
+    """
+    # Build SSH info if this is a remote host (only available for online hosts)
+    ssh_info: SSHInfo | None = None
+
+    is_locked: bool | None = None
+    locked_time: datetime | None = None
+    if isinstance(host, OnlineHostInterface):
+        ssh_connection = host.get_ssh_connection_info()
+        if ssh_connection is not None:
+            user, hostname, port, key_path = ssh_connection
+            ssh_info = SSHInfo(
+                user=user,
+                host=hostname,
+                port=port,
+                key_path=key_path,
+                command=f"ssh -i {key_path} -p {port} {user}@{hostname}",
+            )
+        boot_time = host.get_boot_time()
+        uptime_seconds = host.get_uptime_seconds()
+        resource = host.get_provider_resources()
+        is_locked = host.is_lock_held()
+        # Only fetch locked_time when the lock is held to avoid a redundant
+        # SSH stat command on remote hosts (is_lock_held already checked existence).
+        locked_time = host.get_reported_lock_time() if is_locked else None
+    else:
+        boot_time = None
+        uptime_seconds = None
+        resource = None
+
+    certified_data = host.get_certified_data()
+    host_plugin_data = certified_data.plugin
+    # Always use the certified host_name for consistency between online and offline hosts.
+    # Online hosts would otherwise return the SSH hostname (e.g., "r438.modal.host") via
+    # get_name(), while offline hosts return the friendly name from certified data.
+    host_name = certified_data.host_name
+    # SSH activity is tracked at the host level (host_dir/activity/ssh)
+    ssh_activity = (
+        host.get_reported_activity_time(ActivitySource.SSH) if isinstance(host, OnlineHostInterface) else None
+    )
+    host_details = HostDetails(
+        id=host.id,
+        name=host_name,
+        provider_name=host_ref.provider_name,
+        state=host.get_state() if not is_authentication_failure else HostState.UNAUTHENTICATED,
+        image=certified_data.image,
+        tags={**certified_data.user_tags},
+        boot_time=boot_time,
+        uptime_seconds=uptime_seconds,
+        resource=resource,
+        ssh=ssh_info,
+        snapshots=host.get_snapshots(),
+        is_locked=is_locked,
+        locked_time=locked_time,
+        plugin=host_plugin_data,
+        ssh_activity_time=ssh_activity,
+        failure_reason=certified_data.failure_reason,
+    )
+    return host_details, ssh_activity
+
+
+def _build_agent_details_from_online_agent(
+    agent: AgentInterface,
+    host_details: HostDetails,
+    host: OnlineHostInterface,
+    ssh_activity: datetime | None,
+    field_generators: Mapping[str, Mapping[str, Callable[[AgentInterface, OnlineHostInterface], Any]]],
+) -> AgentDetails:
+    """Build AgentDetails from a live agent on an online host."""
+    # Get activity config from host
+    activity_config = host.get_activity_config()
+
+    # Activity times from file mtimes (per-agent)
+    user_activity = agent.get_reported_activity_time(ActivitySource.USER)
+    agent_activity = agent.get_reported_activity_time(ActivitySource.AGENT)
+
+    # start_time from activity/start file mtime (not the status/start_time file)
+    start_time = agent.get_reported_activity_time(ActivitySource.START)
+
+    # runtime_seconds computed from start_time
+    now = datetime.now(timezone.utc)
+    runtime_seconds = (now - start_time).total_seconds() if start_time else None
+
+    # idle_seconds: include host-level ssh_activity; 0.0 if no activity yet
+    idle_seconds = compute_idle_seconds(user_activity, agent_activity, ssh_activity) or 0.0
+
+    # Compute plugin-specific fields from field generators
+    plugin_data: dict[str, Any] = {}
+    for plugin_name, generators in field_generators.items():
+        plugin_fields: dict[str, Any] = {}
+        for field_name, generator in generators.items():
+            value = generator(agent, host)
+            if value is not None:
+                plugin_fields[field_name] = value
+        if plugin_fields:
+            plugin_data[plugin_name] = plugin_fields
+
+    return AgentDetails(
+        id=agent.id,
+        name=agent.name,
+        type=str(agent.agent_type),
+        command=agent.get_command(),
+        work_dir=agent.work_dir,
+        initial_branch=agent.get_created_branch_name(),
+        create_time=agent.create_time,
+        start_on_boot=agent.get_is_start_on_boot(),
+        state=agent.get_lifecycle_state(),
+        url=agent.get_reported_url(),
+        start_time=start_time,
+        runtime_seconds=runtime_seconds,
+        user_activity_time=user_activity,
+        agent_activity_time=agent_activity,
+        idle_seconds=idle_seconds,
+        idle_mode=activity_config.idle_mode.value,
+        idle_timeout_seconds=activity_config.idle_timeout_seconds,
+        activity_sources=tuple(s.value for s in activity_config.activity_sources),
+        labels=agent.get_labels(),
+        host=host_details,
+        plugin=plugin_data,
+    )
+
+
+def _build_agent_details_from_offline_ref(
+    agent_ref: DiscoveredAgent,
+    host_details: HostDetails,
+) -> AgentDetails:
+    """Build AgentDetails from a discovered agent reference when the host is offline."""
+    create_time = agent_ref.create_time or datetime(1970, 1, 1, tzinfo=timezone.utc)
+    return AgentDetails(
+        id=agent_ref.agent_id,
+        name=agent_ref.agent_name,
+        type=str(agent_ref.agent_type) if agent_ref.agent_type else "unknown",
+        command=agent_ref.command or CommandString(""),
+        work_dir=agent_ref.work_dir or Path("/"),
+        initial_branch=agent_ref.created_branch_name,
+        create_time=create_time,
+        start_on_boot=agent_ref.start_on_boot,
+        state=AgentLifecycleState.STOPPED,
+        url=None,
+        start_time=None,
+        runtime_seconds=None,
+        user_activity_time=None,
+        agent_activity_time=None,
+        idle_seconds=None,
+        idle_mode=None,
+        labels=agent_ref.labels,
+        host=host_details,
+        plugin={},
+    )
 
 
 class ProviderInstanceInterface(MutableModel, ABC):
@@ -224,14 +395,57 @@ class ProviderInstanceInterface(MutableModel, ABC):
         self,
         host_ref: DiscoveredHost,
         agent_refs: Sequence[DiscoveredAgent],
-    ) -> tuple[HostDetails, list[AgentDetails]] | None:
-        """Build HostDetails and AgentDetails for a host in an optimized way for listing.
+        field_generators: Mapping[str, Mapping[str, Callable[[AgentInterface, OnlineHostInterface], Any]]]
+        | None = None,
+    ) -> tuple[HostDetails, list[AgentDetails]]:
+        """Build HostDetails and AgentDetails for a host for listing.
 
+        The default implementation connects to the host to collect data field-by-field.
         Providers can override this to collect all needed data in a single
         operation (e.g., one SSH command) instead of making many individual calls.
-        Returns None to fall back to the default per-field collection.
         """
-        return None
+        is_authentication_failure = False
+        try:
+            host = self.get_host(host_ref.host_id)
+        except HostConnectionError as e:
+            logger.debug("Host {} unreachable, falling back to offline data: {}", host_ref.host_id, e)
+            host = self.to_offline_host(host_ref.host_id)
+            is_authentication_failure = isinstance(e, HostAuthenticationError)
+
+        # Build host details
+        host_details, ssh_activity = _build_host_details_from_host(host, host_ref, is_authentication_failure)
+
+        # Get all agents on this host
+        agents: list[AgentInterface] | None = None
+        if isinstance(host, OnlineHostInterface):
+            agents = host.get_agents()
+
+        # Build AgentDetails for each agent on this host
+        resolved_field_generators = field_generators or {}
+        agent_details_list: list[AgentDetails] = []
+        for agent_ref in agent_refs:
+            agent_details: AgentDetails | None = None
+            if agents is not None and isinstance(host, OnlineHostInterface):
+                # Find the agent in the list for running hosts
+                agent = next((a for a in agents if a.id == agent_ref.agent_id), None)
+                if agent is not None:
+                    agent_details = _build_agent_details_from_online_agent(
+                        agent, host_details, host, ssh_activity, resolved_field_generators
+                    )
+                else:
+                    logger.debug(
+                        "Agent {} not found on host {}, using offline data",
+                        agent_ref.agent_id,
+                        host_ref.host_id,
+                    )
+
+            # If this host is offline, or if we failed to find the agent on the online host
+            if agent_details is None:
+                agent_details = _build_agent_details_from_offline_ref(agent_ref, host_details)
+
+            agent_details_list.append(agent_details)
+
+        return host_details, agent_details_list
 
     # =========================================================================
     # Snapshot Methods
